fix(build): keep node:crypto out of the client bundle (unblock main deploy)
main is red on the build job: every deploy since #2075 fails. The surfaced
error is a red herring —

  [og-pages] [spa-bundle-resolver] poll exhausted ... index.html does not exist yet

The real (masked) failure is a fatal rollup error during the client build:

  services/newsletter-subject-variants.mjs: "createHash" is not exported by
  "__vite-browser-external"

newsletter-subject-variants.mjs is reachable from the SPA client bundle
(services/newsletterPreview.ts -> services/newsletter-content.mjs ->
getVariantStyleDirective). A static NAMED import of node:crypto there makes
rollup fail the whole client build, so dist/index.html is never written and the
og-pages closeBundle poll then exhausts — hence the misleading error.

Fix: split the only node:crypto consumer (assignSubjectVariant, server-only:
send-newsletter + A/B report) into a new server-only module
services/newsletter-subject-assign.mjs. The variants module stays browser-safe;
node:crypto never enters the SPA graph. The hash is byte-identical, so variant
assignment is unchanged (no subscriber re-bucketing, historical recompute
intact). Updated the two script importers and the two unit tests.

Supersedes #2111 (same fix, rebased onto current main where #2108 changed the
assignSubjectVariant signature + renamed the A/B-report importer, which left
#2111 conflicting).
