+++ services/newsletter-subject-variants.mjs
@@ -21,7 +21,13 @@
  * text that could drift.
  */
 
-import { createHash } from 'node:crypto';
+// NOTE: this module is browser-safe by construction — it is reachable from the
+// SPA client bundle (services/newsletterPreview.ts → services/newsletter-content.mjs
+// → getVariantStyleDirective). It must therefore NOT import any `node:` builtin.
+// The lone crypto-dependent helper (`assignSubjectVariant`, server-only: send +
+// A/B report) lives in ./newsletter-subject-assign.mjs so `node:crypto` never
+// leaks into the SPA graph (a static `node:crypto` import here makes rollup fail
+// the whole client build: "createHash is not exported by __vite-browser-external").
 
 /** Lowercase+trim — must match scripts/send-newsletter.mjs normalizeEmail. */
 export function normalizeEmail(raw) {
@@ -93,38 +99,10 @@ export function getSubjectVariant(id) {
  * the losing arm keeps getting traffic and the test stays live). */
 export const DEFAULT_EPSILON = 0.2;
 
-/**
- * Deterministically assign a variant id for (email, campaignId).
- * Stable for the lifetime of a campaign (so a subscriber never flip-flops across
- * the daily cron re-runs of one weekly campaign) and rotates across campaigns
- * (so the same subscriber isn't always in the same bucket).
- *
- * Without a `promotedVariant` the split is uniform (the pure A/B baseline). With
- * one, it is epsilon-greedy: a deterministic (1-epsilon) share is sent the
- * winner, the rest explores uniformly — so auto-promotion biases toward the
- * winner while still measuring every arm.
- *
- * @param {string} email
- * @param {string} campaignId
- * @param {{variants?:Array, promotedVariant?:string|null, epsilon?:number}|Array} [opts]
- *        (an array is accepted for backward compat = the variants list)
- * @returns {string} variant id
- */
-export function assignSubjectVariant(email, campaignId, opts = {}) {
-  const o = Array.isArray(opts) ? { variants: opts } : (opts || {});
-  const { variants = SUBJECT_VARIANTS, promotedVariant = null, epsilon = DEFAULT_EPSILON } = o;
-  const list = variants?.length ? variants : SUBJECT_VARIANTS;
-  const key = `${normalizeEmail(email)}|${String(campaignId || '')}`;
-  const digest = createHash('sha256').update(key).digest();
-  // Independent digest slices: one for the uniform pick, one for the explore/
-  // exploit draw — both deterministic so a subscriber is stable within a campaign.
-  const uniformPick = list[digest.readUInt32BE(0) % list.length].id;
-
-  const promoted = list.find((v) => v.id === promotedVariant);
-  if (!promoted) return uniformPick; // no promotion → pure A/B split
-  const draw = digest.readUInt32BE(4) / 0xffffffff; // [0,1)
-  return draw >= epsilon ? promoted.id : uniformPick; // exploit vs explore
-}
+// `assignSubjectVariant` (the only `node:crypto` consumer) lives in the
+// server-only ./newsletter-subject-assign.mjs to keep this module browser-safe
+// (see the header note). Import it from there in the send / A/B-report scripts
+// and the unit tests.
 
 /**
  * Variant-specific fallback subject for a locale, with safe degradation:
